third_party/blink/web_tests/http/tests/devtools/elements/styles/edit-css-with-source-url-expected.txt

Tests file system project mappings.


Running: testEditingRulesInElementsPanelDoesNotAddSourceURLToOriginalFile
Dumping uiSourceCode content:
#inspected {
          color: red;
      }
      
Dumping matched rules:
[expanded] 
element.style { ()

[expanded] 
#inspected { (foo.css:1 -> foo.css:1:13)
    color: red;

[expanded] 
div { (user agent stylesheet)
    display: block;
    unicode-bidi: isolate;

Editing styles from elements panel:
Styles edited.
Dumping matched rules:
[expanded] 
element.style { ()

[expanded] 
#inspected { (foo.css:1 -> foo.css:1:13)
    color: green;
    : ;

[expanded] 
div { (user agent stylesheet)
    display: block;
    unicode-bidi: isolate;

Dumping uiSourceCode content:
#inspected {
          color: green;
      }
      
